refactor(fuzz): split codec fuzz helpers into modules
